# API_BAJATEL/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bajatel - Proyecto DAW",
        "description": "API REST para gestión de contratos de telecomunicaciones (Fibra, TV, Móvil)",
        "contact": {
            "email": "soporte@bajatel.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
        "/contratos/crear": {
            "post": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Crear nuevo contrato",
                "operationId": "ec479ce5360bf594769a57f23061a85e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "iban",
                                    "calle_y_n",
                                    "ciudad",
                                    "provincia",
                                    "codigo_postal"
                                ],
                                "properties": {
                                    "iban": {
                                        "type": "string",
                                        "maxLength": 29,
                                        "minLength": 29,
                                        "example": "ES9121000418450200051332"
                                    },
                                    "calle_y_n": {
                                        "type": "string",
                                        "example": "Calle Mayor 123"
                                    },
                                    "ciudad": {
                                        "type": "string",
                                        "example": "Madrid"
                                    },
                                    "provincia": {
                                        "type": "string",
                                        "example": "Madrid"
                                    },
                                    "codigo_postal": {
                                        "type": "string",
                                        "example": "28001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contrato creado"
                    },
                    "400": {
                        "description": "Usuario ya tiene contrato"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/contratos/actualizar": {
            "put": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Actualizar contrato existente",
                "operationId": "2c7aa552377d31ad07574afe9388e97a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "iban": {
                                        "type": "string"
                                    },
                                    "calle_y_n": {
                                        "type": "string"
                                    },
                                    "ciudad": {
                                        "type": "string"
                                    },
                                    "provincia": {
                                        "type": "string"
                                    },
                                    "codigo_postal": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contrato actualizado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Contrato no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/contratos/cancelar": {
            "delete": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Cancelar/eliminar contrato",
                "operationId": "113a3452f938b3ac45a1afba327cfd22",
                "responses": {
                    "200": {
                        "description": "Contrato eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Contrato no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/contratos/mostrar": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Mostrar contrato del usuario con servicios",
                "operationId": "1af5ecf5e69ae9118c5ddc00a6349f1b",
                "responses": {
                    "200": {
                        "description": "Contrato con servicios"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Contrato no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum - Usar: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}